requirements.txt
anthropic>=0.40.0
pillow>=10.0.0
mss>=9.0.1
python-dotenv>=1.0.0
apscheduler>=3.10.4
imagehash>=4.3.1
pyinstaller>=6.0.0
